+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * InnoDB Dirty pages ratio query #11
 * InnoDB Waits for checkpoint query #12
 * InnoDB Waits for redolog query #13
+* Table cache query #14
 
 ## v0.3.1 (2021-11-03)
 ### Added